+++ R/plotResource.R
@@ -108,13 +108,13 @@ plotlyResourcePred <- function(object,
 #' @seealso [plotting_functions]
 #' @examples
 #' \donttest{
-#' plotResource(NS_params)
+#' plotResourceLevel(NS_params)
 #'
 #' # Returning the data frame
-#' fr <- plotResource(NS_params, return_data = TRUE)
+#' fr <- plotResourceLevel(NS_params, return_data = TRUE)
 #' str(fr)
 #' }
-plotResource <- function(object, return_data = FALSE)
+plotResourceLevel <- function(object, return_data = FALSE)
 {
     if (is(object, "MizerSim")) {
         params <- object@params