+++ NEWS.md
@@ -1,7 +1,7 @@
 # mizerMR 0.0.3
 
 * `plotSpectra()` and `plotlySpectra()` work with multiple resources.
-* New plotting functions `plotResource()` and `plotResourcePred()`.
+* New plotting functions `plotResourceLevel()` and `plotResourcePred()`.
 * `setMultipleResources()` no longer changes the initial resource abundances 
   unless supplied via `initial_resource`.
 * Fix bug preventing `resource_params()` from changing the resource arrays.